+++ dune/geometry/axisalignedcubegeometry.hh
@@ -304,7 +304,7 @@ namespace Dune {
       return true;
     }
 
-    friend Dune::Transitional::ReferenceElement< ctype, Dim<dim> > referenceElement ( const AxisAlignedCubeGeometry & /* geometry */ )
+    friend typename Dune::ReferenceElements< ctype, dim >::ReferenceElement referenceElement ( const AxisAlignedCubeGeometry & /* geometry */ )
     {
       return ReferenceElements< ctype, dim >::cube();
     }