+++ dune/geometry/referenceelement.hh
@@ -161,27 +161,18 @@ namespace Dune {
        *  Denote by E the i-th subentity of codimension c of the current
        *  reference element. This method returns the GeometryType of E.
        *
-       *  \deprecated After dune-2.6 the return value will no longer be
-       *              a reference but a copy. Code storing pointers
-       *              or references to the result must be adjusted.
-       *
        *  \param[in]  i      number of subentity E (0 <= i < size( c ))
        *  \param[in]  c      codimension of subentity E
        */
-      decltype(auto) type(int i, int c) const
+      GeometryType type(int i, int c) const
       {
         return _impl->type(i,c);
       }
 
 
       /** \brief obtain the type of this reference element
-       *
-       *  \deprecated After dune-2.6 the return value will no longer be
-       *              a reference but a copy. Code storing pointers
-       *              or references to the result must be adjusted.
-       *
        */
-      decltype(auto) type() const
+      GeometryType type() const
       {
         return _impl->type();
       }
@@ -193,14 +184,10 @@ namespace Dune {
        *  reference element. This method returns the coordinates of
        *  the center of gravity of E within the current reference element.
        *
-       *  \deprecated After dune-2.6 the return value will no longer be
-       *              a reference but a copy. Code storing pointers
-       *              or references to the result must be adjusted.
-       *
        *  \param[in]  i   number of subentity E (0 <= i < size( c ))
        *  \param[in]  c   codimension of subentity E
        */
-      decltype(auto) position(int i, int c) const
+      Coordinate position(int i, int c) const
       {
         return _impl->position(i,c);
       }
@@ -249,13 +236,9 @@ namespace Dune {
        *  The integration outer normal is the outer normal whose length coincides
        *  with the face's integration element.
        *
-       *  \deprecated After dune-2.6 the return value will no longer be
-       *              a reference but a copy. Code storing pointers
-       *              or references to the result must be adjusted.
-       *
        *  \param[in]  face  index of the face, whose normal is desired
        */
-      decltype(auto) integrationOuterNormal(int face) const
+      Coordinate integrationOuterNormal(int face) const
       {
         return _impl->integrationOuterNormal(face);
       }